+++ src/NetSonar/App.Logger.cs
@@ -1,9 +1,8 @@
 ﻿using Microsoft.Extensions.Logging;
-using NetSonar.Avalonia.Models;
-using System.Diagnostics;
+using StageKit;
 using System;
+using System.Diagnostics;
 using System.IO;
-using Updatum;
 using ZLogger;
 using ZLogger.Providers;
 
@@ -23,7 +22,7 @@ public partial class App
             {
                 // File name determined by parameters to be rotated
                 options.FilePathSelector = (timestamp, sequenceNumber) =>
-                    Path.Combine(LogsPath, $"{timestamp.ToLocalTime():yyyy-MM}_{sequenceNumber:000}.log");
+                    Path.Combine(ApplicationKit.LogsPath, $"{timestamp.ToLocalTime():yyyy-MM}_{sequenceNumber:000}.log");
 
                 // The period of time for which you want to rotate files at time intervals.
                 options.RollingInterval = RollingInterval.Month;
@@ -51,61 +50,8 @@ public partial class App
         });
 
         Logger = factory.CreateLogger(Software);
+        ApplicationKit.Logger = Logger;
     }
-
-    /// <summary>
-    /// Handles and log the unhandled exception.
-    /// </summary>
-    /// <param name="category"></param>
-    /// <param name="ex"></param>
-    public static void HandleUnhandledException(Exception ex, string category)
-    {
-        try
-        {
-            WriteLine(ex);
-            Logger.ZLogCritical(ex, $"{category}");
-
-            if (category == "Task")
-            {
-                if (ex.Message.Contains("org.freedesktop.DBus.Error.ServiceUnknown")
-                    || ex.Message.Contains("org.freedesktop.DBus.Error.UnknownMethod")) return;
-            }
-
-            if (!IsCrashReport)
-            {
-                var report = new CrashReport(ex, category);
-                CrashReports.Add(report);
-                EntryApplication.LaunchNewInstance($"--crash-report \"{report.Id}\"");
-            }
-        }
-        catch (Exception e)
-        {
-            WriteLine(e);
-        }
-
-        PanicSaveSettings();
-        Environment.Exit(-1);
-    }
-
-    /// <summary>
-    /// Handles exceptions that are safe to continue.
-    /// </summary>
-    /// <param name="ex"></param>
-    /// <param name="category"></param>
-    /// <param name="logLevel"></param>
-    public static void HandleSafeException(Exception ex, string category, LogLevel logLevel = LogLevel.Error)
-    {
-        try
-        {
-            WriteLine(ex);
-            Logger.ZLog(logLevel, ex, $"{category}");
-        }
-        catch (Exception e)
-        {
-            WriteLine(e);
-        }
-    }
-
 
     /// <summary>
     /// Writes the specified string to the console and debug output.